requirements.txt
audioread~=2.1.9
numpy>=1.19.0
paho-mqtt==1.5.0
rhasspy-hermes~=0.6.0
soundfile~=0.10.3

